% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_env.R
\docType{data}
\name{.strobe_env}
\alias{.strobe_env}
\title{Internal environment for package-level state}
\format{
An object of class \code{environment} of length 3.
}
\usage{
.strobe_env
}
\description{
This environment is used to store internal session-specific state of the STROBE derivation flow.
It is not exported and should not be accessed directly by users.
}
\keyword{internal}
